Move prisma into api tests dir
